todo_backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==3.13.0
pytest==7.3.0
pytest-django==4.5.2
pytest-cov==3.0.0
#django-environ==0.8.2
django-environ==0.11.2
django-rest-knox==4.2.0
requests